refactor: share precompile modules
