+++ graft/coreth/params/extras/precompiles.go
@@ -6,7 +6,7 @@ package extras
 import (
 	"encoding/json"
 
-	"github.com/ava-labs/avalanchego/graft/coreth/precompile/modules"
+	"github.com/ava-labs/avalanchego/graft/evm/precompile/modules"
 	"github.com/ava-labs/avalanchego/graft/evm/precompile/precompileconfig"
 )
