+++ graft/subnet-evm/accounts/abi/bind/precompilebind/precompile_module_template.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/ava-labs/avalanchego/graft/evm/precompile/precompileconfig"
 	"github.com/ava-labs/avalanchego/graft/evm/precompile/contract"
-	"github.com/ava-labs/avalanchego/graft/subnet-evm/precompile/modules"
+	"github.com/ava-labs/avalanchego/graft/evm/precompile/modules"
 
 	"github.com/ava-labs/libevm/common"
 )
