+++ graft/subnet-evm/precompile/allowlist/config_test.go
@@ -6,9 +6,9 @@ package allowlist_test
 import (
 	"testing"
 
+	"github.com/ava-labs/avalanchego/graft/evm/precompile/modules"
 	"github.com/ava-labs/avalanchego/graft/subnet-evm/precompile/allowlist"
 	"github.com/ava-labs/avalanchego/graft/subnet-evm/precompile/allowlist/allowlisttest"
-	"github.com/ava-labs/avalanchego/graft/subnet-evm/precompile/modules"
 )
 
 var testModule = modules.Module{
