+++ graft/subnet-evm/precompile/contracts/rewardmanager/module.go
@@ -9,8 +9,8 @@ import (
 	"github.com/ava-labs/libevm/common"
 
 	"github.com/ava-labs/avalanchego/graft/evm/precompile/contract"
+	"github.com/ava-labs/avalanchego/graft/evm/precompile/modules"
 	"github.com/ava-labs/avalanchego/graft/evm/precompile/precompileconfig"
-	"github.com/ava-labs/avalanchego/graft/subnet-evm/precompile/modules"
 )
 
 var _ contract.Configurator = (*configurator)(nil)
